feat: 자동차 이름 부여, 5자 제한 테스트
