+++ README.md
@@ -41,8 +41,8 @@
 * 자동차 이름은 쉼표(,)를 기준으로 구분한다.
 * 자동차 경주 게임을 완료한 후 누가 우승했는지를 알려준다. 우승자는 한명 이상일 수 있다.
 ### 구현할 기능
-* [ ] 자동차에 이름 부여하기
-* [ ] 자동차 이름은 5자를 초과할 수 없음 
+* [x] 자동차에 이름 부여하기 테스트
+* [x] 자동차 이름은 5자를 초과할 수 없음 테스트
 * [ ] 자동차를 출렬할 때 이름도 같이 출력
 * [ ] 자동차 이름을 입력 받을때 쉼표(,)를 기준으로 구분
 * [ ] 경주가 끝난후 우승자를 알려주기
