+++ src/main/java/racinggame/CarRacingApplication.java
@@ -1,4 +1,4 @@
-package step3;
+package racinggame;
 
 public class CarRacingApplication {
     public static void main(String[] args) {
